backend/requirement.txt
fastapi
uvicorn
python-dotenv
pydantic
requests
youtube-transcript-api
openai
tiktoken
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
python-multipart